+++ src/gadgets/uint32.rs
@@ -319,7 +319,7 @@ impl UInt32 {
 
         // Compute the maximum value of the sum so we allocate enough bits for
         // the result
-        let mut max_value = (operands.len() as u64) * (u64::from(u32::max_value()));
+        let mut max_value = (operands.len() as u64) * (u64::from(u32::MAX));
 
         // Keep track of the resulting value
         let mut result_value = Some(0u64);
@@ -417,7 +417,7 @@ mod test {
     use crate::ConstraintSystem;
     use bls12_381::Scalar;
     use ff::Field;
-    use rand_core::{RngCore, SeedableRng};
+    use rand_core::{Rng, SeedableRng};
     use rand_xorshift::XorShiftRng;
 
     #[test]